multi user db add product
